Develop ok

+++ app/src/main/java/com/example/notestakingapp/firebase/FirebaseAuthHandler.java
@@ -3,6 +3,7 @@ package com.example.notestakingapp.firebase;
 import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
+import android.net.Uri;
 import android.util.Log;
 import android.widget.Toast;
 
@@ -10,10 +11,23 @@ import androidx.annotation.NonNull;
 
 import com.example.notestakingapp.ui.MainActivity;
 import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.OnFailureListener;
+import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
+import com.google.firebase.storage.FirebaseStorage;
+import com.google.firebase.storage.StorageReference;
+import com.google.firebase.storage.UploadTask;
+
+import java.io.File;
+import java.util.Objects;
+
+
+
+
+
 
 public class FirebaseAuthHandler {
     public static final String TAG = "EmailPassword";
@@ -32,12 +46,10 @@ public class FirebaseAuthHandler {
                     @Override
                     public void onComplete(@NonNull Task<AuthResult> task) {
                         if (task.isSuccessful()) {
-                            // Sign up success
                             Log.d(TAG, "createUserWithEmail:success");
                             FirebaseUser user = mAuth.getCurrentUser();
                             updateUI(user, context);
                         } else {
-                            // If sign up fails, display a message to the user
                             Log.w(TAG, "createUserWithEmail:failure", task.getException());
                             Toast.makeText(context, "Sign Up failed: " + task.getException().getMessage(),
                                     Toast.LENGTH_LONG).show();
@@ -59,6 +71,7 @@ public class FirebaseAuthHandler {
                             editor.putString("userEmail", email);
                             editor.apply();
                             updateUI(user, context);
+                            FirebaseHandler.syncFromFirebase(context);
                         } else {
                             Log.w(TAG, "signInWithEmail:failure", task.getException());
                             Toast.makeText(context, "Sign In failed: " + task.getException().getMessage(),
@@ -69,14 +82,60 @@ public class FirebaseAuthHandler {
                 });
     }
     public void signOut(Context context) {
-        mAuth.signOut();
-        SharedPreferences sharedPref = context.getSharedPreferences("UserPrefs", Context.MODE_PRIVATE);
-        SharedPreferences.Editor editor = sharedPref.edit();
-        editor.clear();
-        editor.apply();
-        Toast.makeText(context, "Signed Out", Toast.LENGTH_SHORT).show();
-        Log.d(TAG, "signOut:success");
-        updateUI(null, context);
+        String userId = FirebaseAuthHandler.getUserId();
+        if (userId != null) {
+            File dbFile = context.getDatabasePath("note.db");
+            if (dbFile.exists()) {
+                dbFile.delete(); // Xóa file local
+                Log.d(TAG, "Local database file deleted");
+            }
+
+            // Xóa SharedPreferences
+            SharedPreferences sharedPref = context.getSharedPreferences("UserPrefs", Context.MODE_PRIVATE);
+            SharedPreferences.Editor editor = sharedPref.edit();
+            editor.clear();
+            editor.apply();
+
+            mAuth.signOut();
+            Toast.makeText(context, "Signed Out", Toast.LENGTH_SHORT).show();
+            Log.d(TAG, "signOut:success");
+            updateUI(null, context);
+        } else {
+            Toast.makeText(context, "No user signed in", Toast.LENGTH_SHORT).show();
+            Log.w(TAG, "signOut:failure - no user");
+        }
+    }
+
+    public void changePassword(String oldPassword, final String newPassword, final Context context) {
+        FirebaseUser user = mAuth.getCurrentUser();
+        if (user != null) {
+            mAuth.signInWithEmailAndPassword(Objects.requireNonNull(user.getEmail()), oldPassword)
+                    .addOnCompleteListener(new OnCompleteListener<AuthResult>() {
+                        @Override
+                        public void onComplete(@NonNull Task<AuthResult> task) {
+                            if (task.isSuccessful()) {
+                                FirebaseUser currentUser = mAuth.getCurrentUser();
+                                if (currentUser != null) {
+                                    currentUser.updatePassword(newPassword)
+                                            .addOnCompleteListener(new OnCompleteListener<Void>() {
+                                                @Override
+                                                public void onComplete(@NonNull Task<Void> task) {
+                                                    if (task.isSuccessful()) {
+                                                        Toast.makeText(context, "Password updated successfully", Toast.LENGTH_SHORT).show();
+                                                    } else {
+                                                        Toast.makeText(context, "Failed to update password: " + Objects.requireNonNull(task.getException()).getMessage(), Toast.LENGTH_SHORT).show();
+                                                    }
+                                                }
+                                            });
+                                }
+                            } else {
+                                Toast.makeText(context, "Authentication failed: " + Objects.requireNonNull(task.getException()).getMessage(), Toast.LENGTH_SHORT).show();
+                            }
+                        }
+                    });
+        } else {
+            Toast.makeText(context, "User not signed in", Toast.LENGTH_SHORT).show();
+        }
     }
     public static String getUserId() {
         FirebaseUser user = FirebaseAuth.getInstance().getCurrentUser();
@@ -89,15 +148,12 @@ public class FirebaseAuthHandler {
     }
     public void updateUI(FirebaseUser user, Context context) {
         if (user != null) {
-            // save user info in shared preferences
             SharedPreferences preferences = context.getSharedPreferences("UserPrefs", Context.MODE_PRIVATE);
             SharedPreferences.Editor editor = preferences.edit();
             editor.putString("user_id", user.getUid());
             editor.putString("user_email", user.getEmail());
             editor.apply();
             Toast.makeText(context, "Welcome, " + user.getEmail(), Toast.LENGTH_SHORT).show();
-
-            // redirect to home screen
             Intent intent = new Intent(context, MainActivity.class);
             intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK);
             context.startActivity(intent);
@@ -105,6 +161,7 @@ public class FirebaseAuthHandler {
             Toast.makeText(context, "User not signed in", Toast.LENGTH_SHORT).show();
         }
     }
+
     public void reload(Context context) {
         FirebaseUser user = mAuth.getCurrentUser();
         if (user != null) {
@@ -130,3 +187,13 @@ public class FirebaseAuthHandler {
         return null;
     }
 }
+
+
+
+
+
+
+
+
+
+
